feat: create UserController for api resource
